This should be an int64

+++ src/GUI/TracksView.cpp
@@ -600,13 +600,13 @@ QList<AbstractGraphicsMediaItem*> TracksView::mediaItems( const QPoint& pos )
     return mediaCollisionList;
 }
 
-void TracksView::setCursorPos( int pos )
+void TracksView::setCursorPos( qint64 pos )
 {
     if ( pos < 0 ) pos = 0;
     m_cursorLine->setCursorPos( pos );
 }
 
-int TracksView::cursorPos()
+qint64 TracksView::cursorPos()
 {
     return m_cursorLine->cursorPos();
 }
